fix(q9-07/phase-e review): P1+F1+P2 review findings (5 fixes)
Pre-push code review surfaced 5 findings — all addressed before push.

P1 — Decimal wire contract sẽ fail FE happy path
* schemas/admission.py: PreviewPriorityKvResponse + PriorityObjectCatalogItem
  field types Decimal → float. Pydantic v2 jsonable_encoder() serializes
  Decimal as STRING ("0.75"), but FE Zod uses z.number() + TS .toFixed()
  → runtime crash on non-empty catalog (smoke chưa bắt vì dev DB rỗng).
* Float coerces from BE engine Decimal internally; only wire shape changes.
  Verified via REPL: area_bonus=0.75 (JSON number), bonus_points=1.0.

P1 — Casbin DB seed migration missing
* alembic/versions/q9_07_e0c_seed_priority_casbin.py: new migration seeds
  22 casbin_rule rows (5 routes × user/officer/manager/admin/accountant).
* Per dd5m6n7o8p9q precedent: "policy_templates.py alone is not enough"
  — enforcer reads DB at request time. Without seed → 403 every caller.
* Includes v3 (eft) per memory `casbin-insert-must-include-eft` (NULL eft
  → load_policy() crash → 500 every endpoint).
* Idempotent NOT EXISTS pattern from phase3_06_seed_send_magic_link.

F1 — Missing CasbinAuth on 2 new endpoints
* routers/admissions_v2.py: preview_priority_kv + get_priority_object_catalog
  thiếu CasbinAuth dependency. Sibling 8 endpoints có. Defense-in-depth
  gap: accountant inherit officer → có thể fetch KV preview.
* Added CasbinAuth dep; templates extended với 2 ALLOW (OFFICER + BASIC_USER)
  + 2 DENY (ACCOUNTANT) cho preview + catalog. Migration seeds matching rows.

P2 — PriorityAuditLog ORM not registered
* models/__init__.py: added `from .priority_audit import PriorityAuditLog`
  + "PriorityAuditLog" to __all__. Before fix: hasattr(models,
  'PriorityAuditLog') == False, priority_audit_log not in Base.metadata.
* Wave 2 will use models.PriorityAuditLog INSERT for kv_manual_override
  audit row — break without this fix.

P2 — academic_year hardcode 2026
* routers/admissions_v2.py preview fallback ladder: admission_round
  (eager-loaded chain) → profile.academic_year (snapshot column always
  populated per ADM-017) → DEFAULT 2026 + log warning.
* PriorityTab.tsx: pass `profile.academic_year ?? 2026` thay vì hardcode.
* Multi-year cycle (mùa 2027+) sẽ lookup đúng rate/catalog mà không cần
  code change.

Verification
* docker compose exec backend alembic upgrade head → 22 rows seeded
* SELECT casbin_rule WHERE template_id='_q9_07_e0c_priority_seed' → 22 ✓
* test_q9_07_e0_seed_priority_events.py: 26/26 PASS
* test_notification_parity.py: 55/55 PASS (incl. 3 PRIORITY events trong
  EVENT_GROUP_MAPPING per fix `15e2615c`)
* Pydantic REPL: float coerce verified
* hasattr(models, 'PriorityAuditLog') == True

+++ Backend_FastAPI/app/schemas/admission.py
@@ -2089,14 +2089,19 @@ class PreviewPriorityKvResponse(BaseModel):
     requires_manual_override: bool = Field(False)
     reason: Optional[str] = Field(None)
     breakdown: Optional[dict] = Field(None)
-    area_bonus: Optional[Decimal] = Field(None, description="KV bonus điểm (rate * 1.0)")
+    # Wire contract: Decimal → string by default trên Pydantic v2 json mode
+    # (jsonable_encoder).  FE expects number (Zod z.number() + .toFixed()).
+    # Schema field types là `float` để Pydantic auto-coerce Decimal → float
+    # serialize as JSON number. BE engine vẫn dùng Decimal internal — chỉ
+    # response shape thay đổi.
+    area_bonus: Optional[float] = Field(None, description="KV bonus điểm (rate * 1.0)")
 
     # --- UT (Phase E wireframe extension) ---
-    object_bonus_potential: Optional[Decimal] = Field(
+    object_bonus_potential: Optional[float] = Field(
         None,
         description="MAX UT bonus assuming all submitted codes verified — preview only"
     )
-    object_bonus_verified: Optional[Decimal] = Field(
+    object_bonus_verified: Optional[float] = Field(
         None,
         description="MAX UT bonus restricted to verified codes — engine T6 actual"
     )
@@ -2106,7 +2111,7 @@ class PreviewPriorityKvResponse(BaseModel):
     )
 
     # --- Combined total ---
-    total_bonus_potential: Optional[Decimal] = Field(
+    total_bonus_potential: Optional[float] = Field(
         None,
         description="area_bonus + object_bonus_potential — candidate-facing total"
     )
@@ -2127,7 +2132,10 @@ class PriorityObjectCatalogItem(BaseModel):
     group_code: str = Field(..., description="UT1 | UT2 | UT3+")
     sub_code: str = Field(..., description="2-digit numeric, vd '01'..'07'")
     description: str = Field(..., description="VD 'Anh hùng LLVTND, Anh hùng lao động'")
-    bonus_points: Decimal = Field(..., description="Bonus rate cho diện này (vd 2.00 cho UT01)")
+    # Wire contract: float (not Decimal) — Pydantic auto-coerce DB Decimal
+    # → float for JSON number output. FE Zod expects z.number(). See
+    # PreviewPriorityKvResponse note above for rationale.
+    bonus_points: float = Field(..., description="Bonus rate cho diện này (vd 2.00 cho UT01)")
     evidence_doc_type: Optional[str] = Field(None, description="Gợi ý loại minh chứng (vd 'Quyết định phong tặng')")
 
     model_config = ConfigDict(from_attributes=True)
